Remove extra ReserveEssentials class.
